+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Added a summary tab for the flow details page
+
 ## [0.16.0] - 2024-03-13
 ### Added
 - Introduced dataset scheduling settings page to manage automatically launched ingest/transform flows